Fixing exception handling

+++ src/app/FileHandler.java
@@ -1,9 +1,9 @@
 package app;
 
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.file.FileAlreadyExistsException;
 import java.nio.file.Files;
+import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
 
 public class FileHandler {
@@ -22,9 +22,12 @@ public class FileHandler {
 
     public String writeToFile(Path path, String content) {
         try {
+            if (Files.notExists(path)) {
+                throw new NoSuchFileException("File not found");
+            }
             Files.writeString(path, content);
-        } catch (FileNotFoundException ex) {
-            return "File not found";
+        } catch (NoSuchFileException ex) {
+            return ex.getMessage();
         } catch (IOException ex) {
             return "Something wrong " + ex.getMessage();
         }
@@ -34,7 +37,7 @@ public class FileHandler {
     public String readFromFile(String path) {
         try {
             return Files.readString(Path.of(path));
-        } catch (FileNotFoundException ex) {
+        } catch (NoSuchFileException ex) {
             return "File not found";
         } catch (IOException ex) {
             return "Something wrong " + ex.getMessage();
